- fixed Qt linguist translations - Work on palettes and texture precision:   * the object palette images is now created on demand in a lazy way, and     can just point to the static palette image if it is the same   * some palette image allocations/copies/re-creations have been avoided and a bit optimized   * the size of this object palette can be limited to a given size. By default it is as large as the reference static palette(s). The API for this is internal (C++ only).   * the size of the OpenGL texture can be forced to allow a good precision with min/max factors.   * The SetObjectPalette command has been enhanced with "sizex" and "sizey" optional parameters, which correspond to the OpenGL texture size (not the internal image). Moreover the "palette" param is now optional.   * docs updated accordingly   TODO: there is no GUI to set the sizes right now.
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/anatomist/anatomist-free/branches/3.2@35206 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/volrender/object/volrender.cc
@@ -624,7 +624,7 @@ bool VolRender::glMakeTexImage( const ViewState &state,
   // check if colormap is small enough to fit into OpenGL limitations
   GLint mt = 0;
   glGetIntegerv( GL_MAX_PIXEL_MAP_TABLE, &mt );
-  if( dimx > mt )
+  if( (int) dimx > mt )
     dimx = mt;
 
   // cout << "dim palette: " << dimx << endl;
